Started tests

+++ app/src/main/java/com/github/displace/sdp2022/users/OfflineUser.kt
@@ -2,12 +2,18 @@ package com.github.displace.sdp2022.users
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.util.Log
 import com.github.displace.sdp2022.profile.achievements.Achievement
 import com.github.displace.sdp2022.profile.history.History
 import com.github.displace.sdp2022.profile.statistics.Statistic
+import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.builtins.ArraySerializer
 import kotlinx.serialization.decodeFromString
+import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encodeToString
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
 import kotlinx.serialization.json.Json
 import java.io.File
 import java.text.SimpleDateFormat
@@ -18,40 +24,131 @@ private const val ACHIEVEMENT_PATH = "user/partial"
 private const val STATS_PATH = "user/partial"
 private const val FRIEND_LIST_PATH = "user/friends"
 private const val GAME_HISTORY_PATH = "user/game_history"
+private const val RDM_PATH = "user/rdm"
 
-class OfflineUser(private val context: Context) : User {
+class OfflineUser(private val context: Context?, private val debug: Boolean = false) : User {
+    /**
+     * Achievements
+     *
+     * A simple wrapper that allows use to serialize and deserialize list of achievements
+     */
+    @Serializable
+    private data class Achievements(@Serializable(with = AchievementArraySerializer::class) var array: MutableList<Achievement>)
+
+    private object AchievementArraySerializer : KSerializer<MutableList<Achievement>> {
+        override fun deserialize(decoder: Decoder): MutableList<Achievement> {
+            return decoder.decodeSerializableValue(ArraySerializer(Achievement.serializer()))
+                .toMutableList()
+
+        }
+
+        override val descriptor: SerialDescriptor =
+            ArraySerializer(Achievement.serializer()).descriptor
+
+        override fun serialize(encoder: Encoder, value: MutableList<Achievement>) {
+            encoder.encodeSerializableValue(
+                ArraySerializer(Achievement.serializer()),
+                value.toTypedArray()
+            )
+        }
+    }
+
+    /**
+     * Statistics
+     *
+     * A simple wrapper that allows use to serialize and deserialize list of statistics
+     */
+    @Serializable
+    private data class Statistics(@Serializable(with = StatsArraySerializer::class) var array: MutableList<Statistic>)
+
+    private object StatsArraySerializer : KSerializer<MutableList<Statistic>> {
+        override val descriptor: SerialDescriptor =
+            ArraySerializer(Statistic.serializer()).descriptor
+
+        override fun deserialize(decoder: Decoder): MutableList<Statistic> {
+            return ArraySerializer(Statistic.serializer()).deserialize(decoder).toMutableList()
+        }
+
+        override fun serialize(encoder: Encoder, value: MutableList<Statistic>) {
+            ArraySerializer(Statistic.serializer()).serialize(encoder, value.toTypedArray())
+        }
+    }
+
+    /**
+     * History
+     *
+     * A simple wrapper that allows use to serialize and deserialize list of history
+     */
+    @Serializable
+    private data class Histories(@Serializable(with = HistoryArraySerializer::class) var array: MutableList<History>)
+
+    private object HistoryArraySerializer : KSerializer<MutableList<History>> {
+        override fun deserialize(decoder: Decoder): MutableList<History> {
+            return decoder.decodeSerializableValue(ArraySerializer(History.serializer()))
+                .toMutableList()
+        }
+
+        override val descriptor: SerialDescriptor =
+            ArraySerializer(History.serializer()).descriptor
+
+        override fun serialize(encoder: Encoder, value: MutableList<History>) {
+            ArraySerializer(History.serializer()).serialize(encoder, value.toTypedArray())
+        }
+    }
+
+    /**
+     * Friend list
+     *
+     * A simple wrapper that allows use to serialize and deserialize list of friends
+     */
+    @Serializable
+    private data class FriendList(@Serializable(with = FriendArraySerializer::class) var array: MutableList<PartialUser>)
+
+    private object FriendArraySerializer : KSerializer<MutableList<PartialUser>> {
+        override fun deserialize(decoder: Decoder): MutableList<PartialUser> {
+            return decoder.decodeSerializableValue(ArraySerializer(PartialUser.serializer()))
+                .toMutableList()
+        }
+
+        override val descriptor: SerialDescriptor =
+            ArraySerializer(PartialUser.serializer()).descriptor
+
+        override fun serialize(encoder: Encoder, value: MutableList<PartialUser>) {
+            ArraySerializer(PartialUser.serializer()).serialize(encoder, value.toTypedArray())
+        }
+    }
+
+
     private lateinit var partialUser: PartialUser
-
-    private lateinit var achievements: MutableList<Achievement>
-    private lateinit var stats: MutableList<Statistic>
-    private var friendsList: MutableList<PartialUser> = mutableListOf()
-
-    private lateinit var gameHistory: MutableList<History>
+    private lateinit var achievements: Achievements
+    private lateinit var stats: Statistics
+    private lateinit var friendsList: FriendList
+    private lateinit var gameHistory: Histories
 
     init {
         initUser()
         Thread.sleep(3000)
     }
 
-    override fun getPartialUser(): PartialUser {
-        return readFile(PARTIAL_USER_PATH) as PartialUser
+    override fun getPartialUser(): PartialUser? {
+        return readFile(PARTIAL_USER_PATH) as PartialUser?
     }
 
     @Suppress("CAST_NEVER_SUCCEEDS")
     override fun addAchievement(ach: Achievement) {
-        achievements.add(ach)
+        achievements.array.add(ach)
         writeToFile(
             ACHIEVEMENT_PATH,
-            ach as Serializable
+            ach
         )
     }
 
     @Suppress("CAST_NEVER_SUCCEEDS")
     override fun updateStats(statName: String, newValue: Long) {
-        for (i in 0..stats.size) {
-            if (statName == stats[i].name) {
-                stats[i].value = newValue
-                writeToFile(STATS_PATH, stats[i] as Serializable)
+        for (i in 0..stats.array.size) {
+            if (statName == stats.array[i].name) {
+                stats.array[i].value = newValue
+                writeToFile(STATS_PATH, stats.array[i])
                 return
             }
         }
@@ -59,31 +156,31 @@ class OfflineUser(private val context: Context) : User {
 
     @Suppress("CAST_NEVER_SUCCEEDS")
     override fun addFriend(partialU: PartialUser) {
-        if (!containsPartialUser(friendsList, partialU)) {
-            friendsList.add(partialU)
-            writeToFile(FRIEND_LIST_PATH, friendsList as Serializable)
+        if (!containsPartialUser(friendsList.array, partialU)) {
+            friendsList.array.add(partialU)
+            writeToFile(FRIEND_LIST_PATH, friendsList)
         }
     }
 
     @Suppress("CAST_NEVER_SUCCEEDS")
     override fun removeFriend(partialU: PartialUser) {
-        if (containsPartialUser(friendsList, partialU)) {
-            friendsList.remove(partialU)
-            writeToFile(FRIEND_LIST_PATH, friendsList as Serializable)
+        if (containsPartialUser(friendsList.array, partialU)) {
+            friendsList.array.remove(partialU)
+            writeToFile(FRIEND_LIST_PATH, friendsList)
         }
     }
 
     @Suppress("CAST_NEVER_SUCCEEDS")
     override fun addGameInHistory(map: String, date: String, result: String) {
         val history = History(map, date, result)
-        gameHistory.add(history)
-        writeToFile(GAME_HISTORY_PATH, gameHistory as Serializable)
+        gameHistory.array.add(history)
+        writeToFile(GAME_HISTORY_PATH, gameHistory)
     }
 
     @Suppress("CAST_NEVER_SUCCEEDS")
     override fun changeUsername(newName: String) {
         partialUser.username = newName
-        writeToFile(PARTIAL_USER_PATH, partialUser as Serializable)
+        writeToFile(PARTIAL_USER_PATH, partialUser)
     }
 
     override fun removeUserFromDatabase() {
@@ -95,19 +192,19 @@ class OfflineUser(private val context: Context) : User {
     }
 
     override fun getAchievements(): MutableList<Achievement> {
-        return achievements
+        return achievements.array
     }
 
     override fun getStats(): List<Statistic> {
-        return stats
+        return stats.array
     }
 
     override fun getFriendsList(): MutableList<PartialUser> {
-        return friendsList
+        return friendsList.array
     }
 
     override fun getGameHistory(): MutableList<History> {
-        return gameHistory
+        return gameHistory.array
     }
 
     override fun equals(other: Any?): Boolean {
@@ -141,26 +238,34 @@ class OfflineUser(private val context: Context) : User {
     private fun initializeAchievements() {
         val riddenAchievements: MutableList<Achievement>? =
             readFile(ACHIEVEMENT_PATH) as MutableList<Achievement>?
-        achievements = riddenAchievements ?: mutableListOf(
-            Achievement(
-                "Create your account !",
-                getCurrentDate()
+        achievements = Achievements(
+            riddenAchievements ?: mutableListOf(
+                Achievement(
+                    "Create your account !",
+                    getCurrentDate()
+                )
             )
         )
         if (riddenAchievements == null) {
-            writeToFile(ACHIEVEMENT_PATH, achievements as Serializable)
+            writeToFile(ACHIEVEMENT_PATH, achievements)
         }
     }
 
     @Suppress("UNCHECKED_CAST")
     private fun initializeStats() {
         val riddenStats: MutableList<Statistic>? = readFile(STATS_PATH) as MutableList<Statistic>?
-        stats = riddenStats ?: mutableListOf(
-            Statistic(
-                "stat1",
-                0
-            ), Statistic("stat2", 0)
+        stats = Statistics(
+            riddenStats ?: mutableListOf(
+                Statistic(
+                    "stat1",
+                    0
+                ), Statistic("stat2", 0)
+            )
         )
+
+        if (riddenStats == null) {
+            writeToFile(STATS_PATH, stats)
+        }
 
     }
 
@@ -168,23 +273,33 @@ class OfflineUser(private val context: Context) : User {
     private fun initializeFriendsList() {
         val riddenFriendsList: MutableList<PartialUser>? =
             readFile(FRIEND_LIST_PATH) as MutableList<PartialUser>?
-        friendsList = riddenFriendsList ?: mutableListOf(
-            PartialUser(
-                "dummy_friend_username",
-                "dummy_friend_id"
+        friendsList = FriendList(
+            riddenFriendsList ?: mutableListOf(
+                PartialUser(
+                    "dummy_friend_username",
+                    "dummy_friend_id"
+                )
             )
         )
+
+        if (riddenFriendsList == null) {
+            writeToFile(FRIEND_LIST_PATH, friendsList)
+        }
     }
 
     @Suppress("UNCHECKED_CAST")
     private fun initializeGameHistory() {
         val riddenGameHistory: MutableList<History>? =
             readFile(GAME_HISTORY_PATH) as MutableList<History>?
-        gameHistory = riddenGameHistory ?: mutableListOf(
+        gameHistory = Histories(riddenGameHistory ?: mutableListOf(
             History(
                 "dummy_map", getCurrentDate(), "VICTORY"
             )
-        )
+        ))
+
+        if (riddenGameHistory == null) {
+            writeToFile(GAME_HISTORY_PATH, gameHistory)
+        }
 
     }
 
@@ -192,18 +307,25 @@ class OfflineUser(private val context: Context) : User {
     * Utility functions for this class
     */
 
-    private fun writeToFile(path: String, serializable: Serializable) {
+    private fun writeToFile(path: String, serializable: Any) {
+        if (debug)
+            return
+
         //Create cache file in the cachedFile directory
-        File.createTempFile(path, null, context.cacheDir)
+        File.createTempFile(path, null, context?.cacheDir ?: File(RDM_PATH))
 
         //Open the file and write the serializable object
-        val cachedFile = File(context.cacheDir, path)
+        val cachedFile = File(context?.cacheDir ?: File(RDM_PATH), path)
         cachedFile.printWriter().print(Json.encodeToString(serializable))
     }
 
     private fun readFile(path: String): Any? {
+        if (debug)
+            return null
+
+
         //Get the cached file
-        val cachedFile = File(context.cacheDir, path)
+        val cachedFile = File(context?.cacheDir ?: File(RDM_PATH), path)
 
         //Read it
         val jsonFormatObject = cachedFile.bufferedReader().use { it.readText() }
@@ -217,8 +339,12 @@ class OfflineUser(private val context: Context) : User {
     }
 
     private fun deleteContentFile(path: String) {
+        if (debug)
+            return
+
+
         //Get the cached file
-        val cachedFile = File(context.cacheDir, path)
+        val cachedFile = File(context?.cacheDir ?: File(RDM_PATH), path)
 
         //Delete it
         cachedFile.delete()
